requirements.txt
Django==4.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
dnspython==2.3.0
pip==23.0.1
psycopg2==2.8.6
